Added dodgy hostname setter oneshot service
In GNS3 I want a user to be able to change the name of the router, and
that this can change the hostname of the linux server, so when they
access the console they can easily see which machine they're on.

I also wanted to to it without hacking the gns3server code. I couldn't
easily pass the hostname in but I could pass it in via a null device
with the rounter name its id:

    /dev/virtio-ports/frr.router.hostname.%vm-name%

Then write a script that sets the hostname based on the existence of
this device. Make the script a systemd oneshot service to start at boot
and it works.

This means changing the name of the FRR router in the GNS3 interface, all
you need to do is restat the router (stop and start the device) and
it'll apply the name to the router. This saves you having to log in as
root and running hostname yourself.

Also if you name all your FRR routers before turning them on, then it'll
just work.